Added the 2007 referendum on the directors determining the term length
It was a bit messed up as the index.wml pointed to the very same
election twice, i.e. the election_id=1 *was* indeed the 2005-10
referendum. I guess that the plan was, back when the index.wml was
built, to reference the election_id=4, which is the term length
referendum, which didn't have a subfolder back then.