Bylaws were converted to reStructuredText
This translation allows further modifications. We intend to change the
bylaws and this conversion shall facilitate making diffs and theme the
resulting documents more easily, i.e. HTML with CSS or LaTeX.

This is intended to be the conversion of the current bylaws in PDF
format. Unfortunately, the source of the PDF is not available, hence we
needed to convert from the binary PDF.